+++ src/Task.ts
@@ -4,9 +4,12 @@
  * @since 0.1.0
  */
 
-import { Task } from "fp-ts/Task"
+import * as T from "fp-ts/Task"
+type Task<A> = T.Task<A>
 import { IO } from "fp-ts/IO"
+import { Endomorphism } from "fp-ts/Endomorphism"
 import { fieldMilliseconds, Milliseconds, now, unMilliseconds } from "./Date"
+import { when as _when } from "./Applicative"
 
 /**
  * Wait for the specified number of milliseconds before resolving.
@@ -93,3 +96,30 @@ export const elapsed =
  * @since 0.12.0
  */
 export const execute = <A>(x: Task<A>): Promise<A> => x()
+
+/**
+ * Conditional execution of a `Task`. Helpful for things like asychronous
+ * logging.
+ *
+ * @example
+ * import { flow, pipe } from 'fp-ts/function';
+ * import { Predicate } from 'fp-ts/Predicate';
+ * import { when } from 'fp-ts-std/Task';
+ * import * as TE from 'fp-ts/TaskEither';
+ * import * as T from 'fp-ts/Task';
+ * import { log } from 'fp-ts/Console';
+ *
+ * const logAsync = flow(log, T.fromIO);
+ * const isInvalid: Predicate<number> = n => n !== 42;
+ *
+ * pipe(
+ *   TE.of(123),
+ *   TE.chainFirstTaskK(n =>
+ *     when(isInvalid(n))(logAsync(n))),
+ * );
+ *
+ * @since 0.12.0
+ */
+export const when: (x: boolean) => Endomorphism<Task<void>> = _when(
+  T.ApplicativePar,
+)